Increase max inotify instance
Because of the limit to the amount of inotify fd instance (i.e = 128),
total instances of application execution might be limited. Some _bad_
application, moreover, could use a lot of inotify fd instances
abnormally using eglibc API.

In order to prevent this situation, this work sets the max inotify fd
instance to 1024 as defult value.

Change-Id: Ibb2efdf2de3e0ccb2bf92996a375e694358eaf55
Signed-off-by: Sangjung Woo <sangjung.woo@samsung.com>